# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ZapCentral - Multi-tenant WhatsApp customer service platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants/{tenantId}/departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List departments in a tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /tenants/{tenantId}/departments/{departmentId}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get department by ID
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "404":
          description: Not found
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentUpdate"
      responses:
        "200":
          description: Department updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete a department
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tenants/{tenantId}/departments/{departmentId}/agents:
    post:
      operationId: addDepartmentAgent
      tags: [departments]
      summary: Add an agent to a department
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentAgentInput"
      responses:
        "201":
          description: Agent added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentAgent"
  /tenants/{tenantId}/departments/{departmentId}/agents/{userId}:
    delete:
      operationId: removeDepartmentAgent
      tags: [departments]
      summary: Remove an agent from a department
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Agent removed

  # --- Super Admin ---
...
components:
  schemas:
...
    Department:
      type: object
      properties:
        id:
          type: number
        tenantId:
          type: number
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string
        status:
          type: string
          enum: [active, inactive]
        agentCount:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - name
        - color
        - status
        - agentCount
        - createdAt
        - updatedAt
    DepartmentInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        color:
          type: string
      required:
        - name
    DepartmentUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        color:
          type: string
        status:
          type: string
          enum: [active, inactive]
    DepartmentAgent:
      type: object
      properties:
        departmentId:
          type: number
        clerkUserId:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        isPrimary:
          type: boolean
        addedAt:
          type: string
          format: date-time
      required:
        - departmentId
        - clerkUserId
        - email
        - isPrimary
        - addedAt
    DepartmentAgentInput:
      type: object
      properties:
        clerkUserId:
          type: string
        isPrimary:
          type: boolean
      required:
        - clerkUserId
...